refactor ixmp dependency to own file
